Move everything to parse_arguments file, import parseArgs
